indepenant gui app

+++ dmsc/trimesh.h
@@ -10,7 +10,7 @@ namespace trimesh
 
   const uint gc_max_cell_dim =    tri_cc_t::cc_dim;
   typedef tri_cc_t::cellid_t      cellid_t;
-  typedef double                  fn_t;
+  typedef float                   fn_t;
   typedef std::vector<fn_t>       fn_list_t;
 
   typedef boost::numeric::ublas::bounded_vector<int,2>  int_pair_t;
